Add a spinning-cube-plus-frame-counter sanity check before game loading
Fixed the diagnostic entry-counter print's real bug first: cls.disable_screen
is a float, not an int, so the earlier %d format was a variadic type
mismatch that misaligned every argument after it -- scr_initialized/
con.initialized were reading garbage, not real values.

Then, per request, add a much simpler sanity check ahead of the whole
game-loading rabbit hole this investigation has been chasing bugs
through: draw a rotating cube plus an updating frame-counter bar with
the exact same GL1 context right after it's created, using only real
GLES1 core calls (vertex arrays + glDrawArrays -- GLES1 has no
glBegin/glEnd/glVertex3f/GL_QUADS at all, that's desktop-GL-only and
was rejected outright at compile time). kaios_startcmd (demomap/map)
is deferred out of Qcommon_Init() and only fires once OK is pressed,
so a real game/menu/console load is never attempted against a
renderer that hasn't been visually confirmed to present a frame yet.
Gated the same way as the rest of this investigation's diagnostics:
real implementation in gl1_main.c (RENDERER=gl1 only), always-off stub
in sw_main.c so the shipped soft build is unaffected.

+++ engine/src/client/cl_screen.c
@@ -1832,8 +1832,13 @@ SCR_UpdateScreen(void)
 	{
 		static int kaios_scr_entry_count = 0;
 		kaios_scr_entry_count++;
-		Com_Printf("KAIOS_SCR_ENTRY: #%d disable_screen=%d scr_initialized=%d con.initialized=%d\n",
-			kaios_scr_entry_count, cls.disable_screen, scr_initialized, con.initialized);
+		/* cls.disable_screen is a float (client.h), not an int -- an
+		 * earlier version of this print used %d for it, which is a
+		 * variadic type mismatch that also misaligns every argument
+		 * after it, so scr_initialized/con.initialized were reading
+		 * garbage, not their real values. */
+		Com_Printf("KAIOS_SCR_ENTRY: #%d disable_screen=%.0f scr_initialized=%d con.initialized=%d\n",
+			kaios_scr_entry_count, cls.disable_screen, (int)scr_initialized, (int)con.initialized);
 	}
 #endif
